Cluster id type was changed from int to short.

+++ client/src/main/java/com/orientechnologies/orient/client/remote/OStorageRemote.java
@@ -1221,7 +1221,7 @@ public class OStorageRemote extends OStorageAbstract implements OStorageProxy, O
             network.writeInt(-1);
 
           if (network.getSrvProtocolVersion() >= 18)
-            network.writeInt(iRequestedId);
+            network.writeShort((short) iRequestedId);
         } finally {
           endRequest(network);
         }
